% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R
\name{chatr_addin}
\alias{chatr_addin}
\title{ChatR RStudio Add-ins}
\usage{
chatr_addin()
}
\description{
These functions provide RStudio add-in integration for ChatR.
They can be accessed through the RStudio Addins menu.
Launch ChatR Interactive Assistant
}
