% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_addins.R
\name{chatr_smart_complete}
\alias{chatr_smart_complete}
\title{Smart Code Completion Helper}
\usage{
chatr_smart_complete()
}
\description{
This function provides intelligent code suggestions based on context.
It's designed to be called from custom key bindings.
}
